% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromint.R
\docType{package}
\name{chromint}
\alias{chromint}
\alias{chromint-package}
\title{chromint: A package to analyse chromatin interactions.}
\description{
This package provides functions to analyse genome-wide chromatin intraction
data, like Hi-C. There are functions to parse high-resolution Hi-C data from
Rao et al. 2014, Cell to quantify intreaction frequencies between pairs of
genomic regions (like genes and enhancers) and randomizations to compute
enrichment of interactions.
}
