% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRaoHiCtoGI.R
\name{getBinGR}
\alias{getBinGR}
\title{Get bins of a single chromosome.}
\usage{
getBinGR(chr, resolution, seqInfo)
}
\arguments{
\item{chr}{Chromosome name}

\item{resolution}{The bin size resolution in base pairs}

\item{seqInfo}{A seqinfo object holding the length of each chromosome.}
}
\value{
A GRanges object with bins
}
\description{
This function returns a \code{\link[GenomicRanges]{GRanges}} object with bins of a single
chromosome by given bin size.
}
\keyword{internal}

