% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRaoHiCtoGI.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Normalize interaction matrix using normalization vector.}
\usage{
normalizeMatrix(M, v)
}
\arguments{
\item{M}{a numeric \code{\link[Matrix]{Matrix}} with size n*n.}

\item{v}{a numeric vector of size n.}
}
\value{
A \code{\link[Matrix]{Matrix}} with normalized counts.
}
\description{
Given the raw contact matrix \eqn{M \in R^{nxn}} and normalization vecotor \eqn{v \in
R^{n}}, the normalized counts \eqn{M^{*}_{i,j} = M_{i,j} / (v_{i} * v_{j})} or
written as matrix product \eqn{M^{*} = diag(v^{-1}) * M * diag(v^{-1})}.
}
