% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCaptureHiCtoGI.R
\name{parseCaptureHiC}
\alias{parseCaptureHiC}
\title{Parse Capture HI-C data from Mifsud et al. 2015 as \code{GInteractions}
object.}
\usage{
parseCaptureHiC(infile, seqInfo = NULL)
}
\arguments{
\item{infile}{Input file with interactions.}

\item{seqInfo}{A seqinfo object holding the length of each chromosome.}
}
