% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRaoHiCtoGI.R
\name{getChrToBinOffset}
\alias{getChrToBinOffset}
\title{Get bin offset for each chromosome.}
\usage{
getChrToBinOffset(chromosomes, resolution, seqInfo)
}
\arguments{
\item{chromosomes}{Character vecotr with chromosome identifiers.}

\item{resolution}{The bin size resolution in base pairs}

\item{seqInfo}{A seqinfo object holding the length of each chromosome.}
}
\value{
A numeric vector with offset for each chromosome.
}
\description{
This function creates offset of bin indices for each chromosome.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.ensGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.ensGene
offset <- getChrToBinOffset(c("chr1", "chr2"), 50000, seqinfo(txdb))
}
\keyword{internal}

