% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCaptureHiCtoGI.R
\name{parseCaptureHiC}
\alias{parseCaptureHiC}
\title{Parse Capture Hi-C data as \code{GInteractions} object.}
\usage{
parseCaptureHiC(infile, seqInfo = NULL)
}
\arguments{
\item{infile}{Input file with interactions.}

\item{seqInfo}{A seqinfo object holding the length of each chromosome.}
}
\value{
A \code{\link[InteractioSet]{GInteractions}} object
}
\description{
In the paper Mifsud et al. 2015
(\url{https://www.ncbi.nlm.nih.gov/pubmed/25938943}) promoter caputure Hi-C
data is provied for two cell types. This data can be downloaded from the
supplemental material of the studay as .zip archive:
\url{http://www.ebi.ac.uk/arrayexpress/files/E-MTAB-2323/E-MTAB-2323.additional.1.zip}.
Each file in this zip archive can be parsed separately with this function as
an \code{\link[InteractioSet]{GInteractions}} object.
}
