% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRaoHiCtoGI.R
\name{parseRaoHiCtoGI}
\alias{parseRaoHiCtoGI}
\title{Parse Hi-C interactions from Rao et. al 2014 as GInteractions object.}
\usage{
parseRaoHiCtoGI(cell, resolution, baseDir, seqInfo,
  normalizeByExpected = FALSE, mapQualStr = "MAPQGE30", normStr = "KR")
}
\arguments{
\item{cell}{The cell type of the input experiment}

\item{resolution}{The matrix resolution of the Hi-C map in bp}

\item{baseDir}{The path to the directory with the Hi-C data. This should
contain subdirectories for each cell-type}

\item{seqInfo}{A seqinfo object of the genome}

\item{normalizeByExpected}{Should contact frequencies be normalized by
distance. Default is FALSE.}

\item{mapQualStr}{String representing the mapping quality threshold. Defualt
is "MAPQGE30".}

\item{normStr}{String representation of the normalization method. Default is
"KR". This is not implemented yet.}
}
\value{
GInteractions object with all cis- and trans-interactions
}
\description{
This function parses interactions from a single experiment by a given
resolution.
}
\keyword{Hi-C}
